Kraj

+++ backend/controllers/komentari.js
@@ -1,6 +1,7 @@
 const komentariRouter = require('express').Router()
 const Komentar = require('../models/komentar')
 const Korisnik = require('../models/korisnik')
+const autorizacija = require('../utils/autorizacija')
 
 komentariRouter.get('/', async (req, res) => {
     const rezultat = await Komentar.find({})
@@ -19,11 +20,27 @@ komentariRouter.get('/:id', async (req, res) => {
 })
 
 komentariRouter.delete('/:id', async (req, res) => {
+    const token = autorizacija.dohvatiToken(req)
+    const dekodiraniToken = autorizacija.verificirajToken(token)
+    if(!token || !dekodiraniToken.id){
+        return res.status(401).json({error: 'Neispravan token'})
+        //ako nisi autentificiran, nemaš prava
+    }
+    const korisnikId = req.get('korisnikId')
+    const korisnik = await Korisnik.findById(korisnikId)
     await Komentar.findByIdAndRemove(req.params.id)
+    korisnik.objave = korisnik.objave.filter(o => o.id !== req.params.id)
+    await korisnik.save()
     res.status(204).end()
 })
 
 komentariRouter.put('/:id', async (req, res) => {
+    const token = autorizacija.dohvatiToken(req)
+    const dekodiraniToken = autorizacija.verificirajToken(token)
+    if(!token || !dekodiraniToken.id){
+        return res.status(401).json({error: 'Neispravan token'})
+        //ako nisi autentificiran, nemaš prava
+    }
     const podatak = req.body
     const id = req.params.id
   
@@ -40,6 +57,12 @@ komentariRouter.put('/:id', async (req, res) => {
 
 komentariRouter.post('/', async (req, res, next) => {
     const podatak = req.body
+    const token = autorizacija.dohvatiToken(req)
+    const dekodiraniToken = autorizacija.verificirajToken(token)
+    if(!token || !dekodiraniToken.id){
+        return res.status(401).json({error: 'Neispravan token'})
+        //ako nisi autentificiran, nemaš prava
+    }
     const korisnik = await Korisnik.findById(podatak.korisnikId)
 
   
@@ -47,7 +70,7 @@ komentariRouter.post('/', async (req, res, next) => {
         sadrzaj: podatak.sadrzaj,
         datum: new Date().toISOString(),
         ID_objava: podatak.ID_objava,
-        korisnik: korisnik._id
+        korisnik: podatak.korisnikId
     })
     const spremljeniKomentar = await komentar.save()
     await korisnik.save()
